+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2023-11-15
+
 ### Added
 
 - Support to delete downloaded backups after restore operation (k8s only).
@@ -505,7 +507,8 @@ on storage with the following (case sensitive!) options:<!-- Needs Docs -->
 - all storageClass parameters options are now case insenstive
 - non-debug logging is less verbose in general
 
-[Unreleased]: https://github.com/piraeusdatastore/linstor-csi/compare/v1.2.3...HEAD
+[Unreleased]: https://github.com/piraeusdatastore/linstor-csi/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/piraeusdatastore/linstor-csi/compare/v1.2.3...v1.3.0
 [1.2.3]: https://github.com/piraeusdatastore/linstor-csi/compare/v1.2.2...v1.2.3
 [1.2.2]: https://github.com/piraeusdatastore/linstor-csi/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/piraeusdatastore/linstor-csi/compare/v1.2.0...v1.2.1